discord/requirements.txt
discord.py==2.0.1
kafka-python==2.0.2
